+++ CHANGELOG.md
@@ -11,8 +11,8 @@ Changelog follow https://keepachangelog.com/ format.
 *   `etree`:
     *   Added: `etree.stack` to stack/batch multiple trees of arrays together.
 *   `enp`:
-    *   Added: `enp.batch_dot`: Always dot product on the last axis with broadcasting
-        support (while `np.dot` is inconsistent 1-D vs 2-D).
+    *   Added: `enp.batch_dot`: Always dot product on the last axis with
+        broadcasting support (while `np.dot` is inconsistent 1-D vs 2-D).
     *   Added: `enp.angle_between` to compute angle between 2 n-dimensions
         vectors.
     *   Changed: `enp.project_onto_vector`, `enp.project_onto_plane` supports