+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+<!-- --------- Do not add anything bellow this line --------- -->
+
+## [1.0.0] - 2023-01-09
+
 *   `etree`:
     *   Added: `etree.stack` to stack/batch multiple trees of arrays together.
     *   Added: `etils.etree.py` backend (pure Python implementation to avoid
@@ -38,8 +42,6 @@ Changelog follow https://keepachangelog.com/ format.
     *   Fixed: `local_path.copy('gs://')` uses the correct backend.
 *   All:
     *   Changed: Better error message when missing import.
-
-<!-- --------- Do not add anything bellow this line --------- -->
 
 ## [0.9.0] - 2022-10-28
 
@@ -186,7 +188,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/google/etils/compare/v0.9.0...v1.0.0
 [0.9.0]: https://github.com/google/etils/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/google/etils/compare/v0.7.0...v0.8.0
 [0.7.1]: https://github.com/google/etils/compare/v0.7.0...v0.7.1